+++ src/boilerplate/scenes/mainScene.ts
@@ -7,6 +7,12 @@
 export class MainScene extends Phaser.Scene {
   private phaserSprite: Phaser.GameObjects.Sprite;
 
+  constructor() {
+    super({
+      key: "MainScene"
+    });
+  }
+
   preload(): void {
     this.load.image("logo", "../assets/boilerplate/phaser.png");
   }